requirements.txt
bardapi==0.1.11
cvlib==0.2.7
googletrans==4.0.0rc1
gTTS==2.3.1
gTTS==2.3.2
opencv_contrib_python==4.6.0.66
opencv_python==4.5.5.64
opencv_python==4.7.0.72
pandas==1.3.4
pandas==1.5.0
Pillow==8.2.0
Pillow==8.4.0
Pillow==9.5.0
Pillow==9.5.0
Requests==2.31.0
streamlit==1.12.0
transformers==4.30.1
